Bug 1977854 - Wasm custom page sizes part x: update wast to 0.225.0 from 0.219.0
This is needed for custom page sizes proposal abbreviation syntax for memories.
Updating this removes the need for the unicode-width patch, so that is removed
as well. Also updates wasm-smith dependency in gluesmith in order to avoid
dependency duplication.

This change also affects how some custom page size tests are parsed, which is
why this commit also modifies some of those tests.